Rename aws_pd -> aws_ebs
