+++ pkg/volume/aws_ebs/aws_pd.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package aws_pd
+package aws_ebs
 
 import (
 	"fmt"
